+++ server/src/main/java/com/codesquad/server/domain/repository/CardRepository.java
@@ -1,5 +1,6 @@
-package com.codesquad.server.domain.card;
+package com.codesquad.server.domain.repository;
 
+import com.codesquad.server.domain.entity.Card;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CardRepository extends CrudRepository<Card, Long> {
