--- server/src/main/java/com/codesquad/server/domain/response/Response.java
@@ -1,14 +1,0 @@
-package com.codesquad.server.domain.response;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class Response {
-    private String message;
-
-    private Object data;
-}
