+++ server/src/main/java/com/codesquad/server/domain/service/UserService.java
@@ -1,8 +1,9 @@
-package com.codesquad.server.application.user;
+package com.codesquad.server.domain.service;
 
-import com.codesquad.server.domain.user.User;
+import com.codesquad.server.domain.entity.User;
 
 public interface UserService {
     public User signUp(User user);
+
     public User signIn(User user);
 }
