+++ server/src/main/java/com/codesquad/server/domain/value/Response.java
@@ -1,0 +1,16 @@
+package com.codesquad.server.domain.value;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class Response {
+    private String message;
+
+    private Object data;
+}
